[roles] Elaborate on how to install & use multiple roles
As already mentioned in the README, only Ansible v2.7 is supported, which
does not understand the concept of multiple roles in one repository. Not
only the meta.yml is needed, but also a workaround on the consumer-side is
required to actually "discover" the roles. This commit elaborates on how
to configure this.