[role:sft-server] Allow to disable SFT metrics and the resulting installation of dependencies
Default: enabled

In some cases, e.g. to test deploy the SFT, metrics are not needed. Feature-flagging this
part simplifies installation a/o updating of sftd.
Allowing to disable the installation of metrics also allows to not require providing a
source (or credentials) to (access) the certificate that is used for Prometheus
authentication.