roles/srv-announcer: simplify fetching
There's few value in providing three different ways to retrieve the srv
announcer. In fact, it's debatable on whether this should belong in this
ansible role at all, given it's not something necessary to deploy SFT,
but something very specific to our environment (and the lack of service
discovery).

Simplify this, by only accepting
```
 - srv_announcer_artifact_file_url
 - srv_announcer_artifact_checksum
```

Which needs to be set from the outside while instantiating this role.

`srv_announcer_artifact_checksum` now is also piped directly to
`get_url.checksum`, meaning it'll need to have the hashing algorithm and
a colon in front.