[roles:sft-server] Add Molecule scenario 'without-metrics'
This change set add a new scenario to test for when metrics installation
is not enabled via inventory variable.
It furthermore introduces a 'verify' step and starts to utilize the test.yml
tasks that already exist. And, another test is being added, that checks if
node metrics are being exposed.